a0 f1

+++ 0x0C-more_malloc_free/100-realloc.c
@@ -33,7 +33,7 @@ return (NULL);
 temp = malloc(new_size);
 if (temp == NULL)
 return (NULL);
-for (i = 0; i < (old_size < new_size ? old_size : new_size); i++)
+for (i = 0; (unsigned int)i < (old_size < new_size ? old_size : new_size); i++)
 ((char *)temp)[i] = ((char *)ptr)[i];
 return (ptr);
 }
